ADK hackathon/agents/api_builders/requirements.txt
protobuf~=6.31.1
pyyaml~=6.0.2
openapi-spec-validator~=0.7.2
python-gitlab~=6.0.0
openapi-generator-cli~=7.13.0
google-adk~=1.4.2